# internal/auth/infra/authhttp/handlers_usuarios.go
package authhttp

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"

	"github.com/abdimuy/msp-api/internal/auth"
	"github.com/abdimuy/msp-api/internal/auth/app"
	"github.com/abdimuy/msp-api/internal/auth/ports/outbound"
	"github.com/abdimuy/msp-api/internal/platform/apperror"
	"github.com/abdimuy/msp-api/internal/platform/pagination"
	"github.com/abdimuy/msp-api/internal/platform/response"
	"github.com/abdimuy/msp-api/internal/platform/validator"
)

// ListarUsuarios handles GET /v2/usuarios.
func (h *Handlers) ListarUsuarios(w http.ResponseWriter, r *http.Request) {
	p, err := pagination.FromRequest(r)
	if err != nil {
		response.Error(w, r, apperror.NewValidation("invalid_pagination", "parámetros de paginación inválidos").WithError(err))
		return
	}
	page, err := h.svc.Listar(r.Context(), outbound.ListParams{Cursor: p.After, PageSize: p.Limit})
	if err != nil {
		response.Error(w, r, err)
		return
	}

	items := make([]UsuarioResponse, 0, len(page.Items))
	for _, u := range page.Items {
		items = append(items, toUsuarioResponse(u))
	}
	response.JSON(w, r, http.StatusOK, ListResponse[UsuarioResponse]{Items: items, NextCursor: page.NextCursor})
}

// ObtenerUsuario handles GET /v2/usuarios/{id}.
func (h *Handlers) ObtenerUsuario(w http.ResponseWriter, r *http.Request) {
	id, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	u, err := h.svc.Obtener(r.Context(), id)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	response.JSON(w, r, http.StatusOK, toUsuarioResponse(u))
}

// CrearUsuario handles POST /v2/usuarios. The caller supplies the
// firebase_uid of an account it already created in Firebase Auth; the row
// ends up as ESTATUS = FIREBASE_USER with that uid bound.
//
// Two outcomes share the 201: a brand-new row, or an existing active
// VENDEDOR_ONLY row (minted from the phone when a cobrador named the person
// as vendedor) promoted in place — same id, real nombre. The office cannot
// tell them apart and does not need to. See app.Service.Crear.
//
// Every collision the endpoint refuses to resolve — the email held by a
// FIREBASE_USER row or by a deactivated one, or the firebase_uid already
// bound elsewhere — surfaces as one 409 with code "usuario_ya_existe".
func (h *Handlers) CrearUsuario(w http.ResponseWriter, r *http.Request) {
	var req CrearUsuarioRequest
	if err := decodeJSON(r, &req); err != nil {
		response.Error(w, r, err)
		return
	}
	if fe := validator.Default().Struct(req); fe != nil {
		response.ValidationError(w, r, fe)
		return
	}
	by, ok := currentUserID(r)
	if !ok {
		response.Error(w, r, apperror.NewUnauthorized("unauthenticated", "no autenticado"))
		return
	}

	u, err := h.svc.Crear(r.Context(), app.CrearParams{
		FirebaseUID: req.FirebaseUID,
		Email:       req.Email,
		Nombre:      req.Nombre,
		Telefono:    req.Telefono,
	}, by)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	response.JSON(w, r, http.StatusCreated, toUsuarioResponse(u))
}

// ActualizarUsuario handles PATCH /v2/usuarios/{id}.
func (h *Handlers) ActualizarUsuario(w http.ResponseWriter, r *http.Request) {
	id, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	var req ActualizarUsuarioRequest
	if decErr := decodeJSON(r, &req); decErr != nil {
		response.Error(w, r, decErr)
		return
	}
	if fe := validator.Default().Struct(req); fe != nil {
		response.ValidationError(w, r, fe)
		return
	}
	by, ok := currentUserID(r)
	if !ok {
		response.Error(w, r, apperror.NewUnauthorized("unauthenticated", "no autenticado"))
		return
	}

	u, err := h.svc.Actualizar(r.Context(), app.ActualizarParams{
		ID:        id,
		Email:     req.Email,
		Nombre:    req.Nombre,
		Telefono:  req.Telefono,
		AlmacenID: req.AlmacenID,
	}, by)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	response.JSON(w, r, http.StatusOK, toUsuarioResponse(u))
}

// DesactivarUsuario handles DELETE /v2/usuarios/{id}.
func (h *Handlers) DesactivarUsuario(w http.ResponseWriter, r *http.Request) {
	id, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	by, ok := currentUserID(r)
	if !ok {
		response.Error(w, r, apperror.NewUnauthorized("unauthenticated", "no autenticado"))
		return
	}
	if err := h.svc.Desactivar(r.Context(), id, by); err != nil {
		response.Error(w, r, err)
		return
	}
	response.NoContent(w)
}

// AsignarRolAUsuario handles POST /v2/usuarios/{id}/roles.
func (h *Handlers) AsignarRolAUsuario(w http.ResponseWriter, r *http.Request) {
	usuarioID, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	var req AsignarRolRequest
	if decErr := decodeJSON(r, &req); decErr != nil {
		response.Error(w, r, decErr)
		return
	}
	if fe := validator.Default().Struct(req); fe != nil {
		response.ValidationError(w, r, fe)
		return
	}
	rolID, parseErr := uuid.Parse(req.RolID)
	if parseErr != nil {
		response.Error(w, r,
			apperror.NewValidation("invalid_uuid", "el rol_id no es un UUID válido").WithError(parseErr),
		)
		return
	}
	by, ok := currentUserID(r)
	if !ok {
		response.Error(w, r, apperror.NewUnauthorized("unauthenticated", "no autenticado"))
		return
	}
	if err := h.svc.AsignarRolAUsuario(r.Context(), usuarioID, rolID, by); err != nil {
		response.Error(w, r, err)
		return
	}
	response.NoContent(w)
}

// ObtenerRolesDeUsuario handles GET /v2/usuarios/{id}/roles.
func (h *Handlers) ObtenerRolesDeUsuario(w http.ResponseWriter, r *http.Request) {
	id, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	roles, err := h.svc.RolesDeUsuario(r.Context(), id)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	items := make([]RolResponse, 0, len(roles))
	for _, rol := range roles {
		items = append(items, toRolResponse(rol))
	}
	response.JSON(w, r, http.StatusOK, ListResponse[RolResponse]{Items: items})
}

// ObtenerPermisosDeUsuario handles GET /v2/usuarios/{id}/permisos. It returns
// the effective union of permisos granted to the usuario via every rol it
// has assigned.
func (h *Handlers) ObtenerPermisosDeUsuario(w http.ResponseWriter, r *http.Request) {
	id, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	permisos, err := h.svc.PermisosEfectivosDeUsuario(r.Context(), id)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	items := make([]PermisoResponse, 0, len(permisos))
	for _, p := range permisos {
		items = append(items, toPermisoResponse(p))
	}
	response.JSON(w, r, http.StatusOK, ListResponse[PermisoResponse]{Items: items})
}

// RevocarRolDeUsuario handles DELETE /v2/usuarios/{id}/roles/{rol_id}.
func (h *Handlers) RevocarRolDeUsuario(w http.ResponseWriter, r *http.Request) {
	usuarioID, err := parseUUIDParam(r, "id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	rolID, err := parseUUIDParam(r, "rol_id")
	if err != nil {
		response.Error(w, r, err)
		return
	}
	if err := h.svc.RevocarRolDeUsuario(r.Context(), usuarioID, rolID); err != nil {
		response.Error(w, r, err)
		return
	}
	response.NoContent(w)
}

// EnsureVendedoresByEmail handles POST /v2/usuarios/ensure-vendedores-by-email.
// The authenticated caller (a cobrador, in practice) submits a list of vendedor
// emails; the service lazily upserts missing ones as VENDEDOR_ONLY usuarios
// and returns the email→usuario_id mapping. No specific permission is
// required: the endpoint is idempotent and side-effect-free as far as the
// caller is concerned.
func (h *Handlers) EnsureVendedoresByEmail(w http.ResponseWriter, r *http.Request) {
	var req EnsureVendedoresByEmailRequest
	if decErr := decodeJSON(r, &req); decErr != nil {
		response.Error(w, r, decErr)
		return
	}
	if fe := validator.Default().Struct(req); fe != nil {
		response.ValidationError(w, r, fe)
		return
	}
	by, ok := currentUserID(r)
	if !ok {
		response.Error(w, r, apperror.NewUnauthorized("unauthenticated", "no autenticado"))
		return
	}

	results, err := h.svc.EnsureVendedoresByEmail(r.Context(), req.Emails, by)
	if err != nil {
		response.Error(w, r, err)
		return
	}
	response.JSON(w, r, http.StatusOK, toEnsureVendedoresResponse(results))
}

// parseUUIDParam reads a URL parameter and parses it as a UUID, returning a
// 422 apperror with a stable "invalid_uuid" code on failure.
func parseUUIDParam(r *http.Request, name string) (uuid.UUID, error) {
	raw := chi.URLParam(r, name)
	id, err := uuid.Parse(raw)
	if err != nil {
		return uuid.Nil, apperror.NewValidation("invalid_uuid", "el identificador en la URL no es un UUID válido").
			WithField("param", name).
			WithError(err)
	}
	return id, nil
}

// currentUserID returns the planted user's ID for use as the "by" actor on
// write commands. The ok==false path should never happen inside the
// authenticated subgroup but is checked defensively.
func currentUserID(r *http.Request) (uuid.UUID, bool) {
	cu, ok := auth.CurrentUserFromContext(r.Context())
	if !ok {
		return uuid.Nil, false
	}
	return cu.ID, true
}